Added file
The main objective of this project is to perform classification on the breast cancer dataset based on
the proteomic expression dataset to identify the subtype of breast cancer. A collection of proteins produced
by cancer cells is known as the breast cancer proteome. I am examining if the cancer subtype can be correctly
identified using only the proteome data for each patient. The target variable in the study is PAM50 mRNA
which is used in breast cancer intrinsic subtyping based on gene expression. The variable is categorical
presenting four subtypes (basal-like, HER2-enriched, Luminal A, and Luminal B). PAMA50 mRNA proteins
are my predictor variables. Using supervised ML classification, I am examining if the cancer subtype can be
correctly identified using only the proteome data for each patient.